+++ composer.json
@@ -12,6 +12,7 @@
         "doctrine/orm": "^2.15",
         "phpdocumentor/reflection-docblock": "^5.3",
         "phpstan/phpdoc-parser": "^1.22",
+        "stof/doctrine-extensions-bundle": "^1.7",
         "symfony/asset": "6.3.*",
         "symfony/console": "6.3.*",
         "symfony/doctrine-messenger": "6.3.*",
