+++ src/Repository/BreederRepository.php
@@ -38,29 +38,17 @@ class BreederRepository extends ServiceEntityRepository
             $this->getEntityManager()->flush();
         }
     }
-
-//    /**
-//     * @return Breeder[] Returns an array of Breeder objects
-//     */
-//    public function findByExampleField($value): array
-//    {
-//        return $this->createQueryBuilder('b')
-//            ->andWhere('b.exampleField = :val')
-//            ->setParameter('val', $value)
-//            ->orderBy('b.id', 'ASC')
-//            ->setMaxResults(10)
-//            ->getQuery()
-//            ->getResult()
-//        ;
-//    }
-
-//    public function findOneBySomeField($value): ?Breeder
-//    {
-//        return $this->createQueryBuilder('b')
-//            ->andWhere('b.exampleField = :val')
-//            ->setParameter('val', $value)
-//            ->getQuery()
-//            ->getOneOrNullResult()
-//        ;
-//    }
+    public function findForHome():array
+    {
+        return $this->createQueryBuilder('b')
+        ->select([
+            'b',
+        ])
+            ->leftJoin('b.offers', 'o')
+            ->orderBy('o.dateTime', 'DESC')
+            ->groupBy('b.id')
+            ->getQuery()
+            ->getResult()
+        ;
+    }
 }
