+++ templates/default/index.html.twig
@@ -5,29 +5,42 @@
 {% block body %}
 <h2 class="mb-2">Nos dernières annonces</h2>
     {% for offer in offers %}
-
-    <div class="card mb-3 w-50 mx-auto" >
-        <div class="d-flex p-2 gap-2 justify-content-center align-items-center"  style="height: 500px;">
+        <div class="card mb-3 w-75 mx-auto" >
+            <div class="d-flex p-2 gap-2 justify-content-center align-items-center" style="height: 400px;">
                 <img src="{{asset('/media/poppy_1.jpg')}}" class="card-img-top rounded h-100 object-fit-cover" alt="...">
                 <img src="{{asset('/media/poppy_2.jpg')}}" class="card-img-top rounded h-100 object-fit-cover" alt="...">
                 <img src="{{asset('/media/poppy_3.jpg')}}" class="card-img-top rounded h-100 object-fit-cover" alt="...">
-        </div>
-        <div class="card-body d-flex align-items-center justify-content-center">
-            <div class="w-50">
-            <h5 class="card-title">{{ offer.title }}</h5>
-
-    {{ offer.dogs|length }}
-            {% for breed in offer.breeds %}
-            {{ breed.name }}
-    {% endfor %}
-            <p class="card-text"><small class="text-body-secondary">Mise à jour le {{ offer.dateTime|date("d/m/Y") }} </small></p>
             </div>
-
-            <a href="" class="btn btn-primary">Voir l'annonce</a>
+            <div class="card-body d-flex align-items-center justify-content-center">
+                <div class="w-75">
+                    <h5 class="card-title">{{ offer.title }}</h5>
+                    <p>Nombre de chiens : {{ offer.dogs|length }}</p>
+                    <p>Races :
+                        {% for breed in offer.breeds %}
+                            <span class="btn btn-light">{{ breed.name }}</span> 
+                        {% endfor %}
+                    </p>
+                    <p class="card-text"><small class="text-body-secondary">Mise à jour le {{ offer.dateTime|date("d/m/Y") }} </small></p>
+                </div>
+                <a href="" class="btn btn-primary">Voir l'annonce</a>
+            </div>
         </div>
-    </div>
     {% endfor %}
 
 <h2>Nos éleveurs et associations</h2>
+<table class="container table table-striped">
+    <tr>
+        <th>Nom de l'éleveur</th>
+        <th>Nombre d'annonces en cours</th>
+        <th>Nombre d'annonces pourvues</th>
+    </tr>
+    {% for breeder in breeders %}
+    <tr>
+        <td> {{ breeder.name }} </td>
+        <td> {{ breeder.offers|length }} </td>
+        <td> {{ breeder.name }} </td>
+    </tr>
+    {% endfor %}
+</table>
 
 {% endblock body %}
